+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ odd bug fix may not get mentioned here at all.
 
 * 2.0.0:
 
+  - Breaking distribution changes:
+
+    + The standalone-optimized tree has been removed. Wed now ships a
+      ``standalone`` tree, which is the same as it has ever been, and a
+      ``packed`` tree (introduced in 1.0.0) that is wed optimized through
+      Webpack.
+
   - Breaking API changes:
 
     + ``Key.__cache`` is no longer public. It was a mistake that it was.